third_party/WebKit/LayoutTests/platform/win/external/wpt/referrer-policy/generic/subresource-test/link-navigate-expected.txt

This is a testharness.js-based test.
PASS Link is responding with HTTP headers 
Harness: the test ran to completion.
